Hardcode issue type of "TASK" whenever we are creating a jira issue Remove all code related to selecting jira issue type ID. Update all relavant queries and mutations
use domain/browse/issue_id for jira external link

Resolve all changes requested from PR
