+++ backend/private-graph/graph/resolver.go
@@ -2659,7 +2659,6 @@ func (r *Resolver) CreateJiraTaskAndAttachment(
 	issueTitle string,
 	issueDescription string,
 	projectId string,
-	issueTypeId string,
 ) error {
 	accessToken, err := r.IntegrationsClient.GetWorkspaceAccessToken(ctx, workspace, modelInputs.IntegrationTypeJira)
 
@@ -2675,7 +2674,7 @@ func (r *Resolver) CreateJiraTaskAndAttachment(
 		Description: issueDescription,
 		Summary:     issueTitle,
 		Project:     jira.JiraIssueProjectData{Id: projectId},
-		IssueType:   jira.JiraIssueTypeData{Id: issueTypeId},
+		IssueType:   jira.JiraIssueTypeData{Id: "10001"},
 	}
 
 	jiraCreateIssueData := jira.JiraCreateIssuePayload{
