+++ backend/private-graph/graph/schema.graphqls
@@ -2092,8 +2092,6 @@ type Mutation {
 		issue_title: String
 		issue_description: String
 		issue_team_id: String
-		issue_project_id: String
-		issue_type_id: String
 		integrations: [IntegrationType]!
 		tags: [SessionCommentTagInput]!
 		additional_context: String
@@ -2108,8 +2106,6 @@ type Mutation {
 		issue_title: String
 		issue_description: String
 		issue_team_id: String
-		issue_project_id: String
-		issue_type_id: String
 		integrations: [IntegrationType]!
 	): SessionComment
 	deleteSessionComment(id: ID!): Boolean
@@ -2134,8 +2130,6 @@ type Mutation {
 		issue_title: String
 		issue_description: String
 		issue_team_id: String
-		issue_project_id: String
-		issue_type_id: String
 		integrations: [IntegrationType]!
 	): ErrorComment
 	removeErrorIssue(error_issue_id: ID!): Boolean
@@ -2149,8 +2143,6 @@ type Mutation {
 		issue_title: String
 		issue_description: String
 		issue_team_id: String
-		issue_project_id: String
-		issue_type_id: String
 		integrations: [IntegrationType]!
 	): ErrorComment
 	deleteErrorComment(id: ID!): Boolean
