+++ backend/private-graph/graph/schema.resolvers.go
@@ -1393,7 +1393,7 @@ func (r *mutationResolver) SaveBillingPlan(ctx context.Context, workspaceID int,
 }
 
 // CreateSessionComment is the resolver for the createSessionComment field.
-func (r *mutationResolver) CreateSessionComment(ctx context.Context, projectID int, sessionSecureID string, sessionTimestamp int, text string, textForEmail string, xCoordinate float64, yCoordinate float64, taggedAdmins []*modelInputs.SanitizedAdminInput, taggedSlackUsers []*modelInputs.SanitizedSlackChannelInput, sessionURL string, time float64, authorName string, sessionImage *string, issueTitle *string, issueDescription *string, issueTeamID *string, issueProjectID *string, issueTypeID *string, integrations []*modelInputs.IntegrationType, tags []*modelInputs.SessionCommentTagInput, additionalContext *string) (*model.SessionComment, error) {
+func (r *mutationResolver) CreateSessionComment(ctx context.Context, projectID int, sessionSecureID string, sessionTimestamp int, text string, textForEmail string, xCoordinate float64, yCoordinate float64, taggedAdmins []*modelInputs.SanitizedAdminInput, taggedSlackUsers []*modelInputs.SanitizedSlackChannelInput, sessionURL string, time float64, authorName string, sessionImage *string, issueTitle *string, issueDescription *string, issueTeamID *string, integrations []*modelInputs.IntegrationType, tags []*modelInputs.SessionCommentTagInput, additionalContext *string) (*model.SessionComment, error) {
 	admin, isGuest := r.getCurrentAdminOrGuest(ctx)
 
 	// All viewers can leave a comment, including guests
@@ -1620,8 +1620,7 @@ func (r *mutationResolver) CreateSessionComment(ctx context.Context, projectID i
 				attachment,
 				title,
 				desc,
-				*issueProjectID,
-				*issueTypeID,
+				*issueTeamID,
 			); err != nil {
 				return nil, e.Wrap(err, "error creating Jira task")
 			}
@@ -1649,7 +1648,7 @@ func (r *mutationResolver) CreateSessionComment(ctx context.Context, projectID i
 }
 
 // CreateIssueForSessionComment is the resolver for the createIssueForSessionComment field.
-func (r *mutationResolver) CreateIssueForSessionComment(ctx context.Context, projectID int, sessionURL string, sessionCommentID int, authorName string, textForAttachment string, time float64, issueTitle *string, issueDescription *string, issueTeamID *string, issueProjectID *string, issueTypeID *string, integrations []*modelInputs.IntegrationType) (*model.SessionComment, error) {
+func (r *mutationResolver) CreateIssueForSessionComment(ctx context.Context, projectID int, sessionURL string, sessionCommentID int, authorName string, textForAttachment string, time float64, issueTitle *string, issueDescription *string, issueTeamID *string, integrations []*modelInputs.IntegrationType) (*model.SessionComment, error) {
 	var project model.Project
 	if err := r.DB.Where("id = ?", projectID).Take(&project).Error; err != nil {
 		return nil, err
@@ -1702,7 +1701,7 @@ func (r *mutationResolver) CreateIssueForSessionComment(ctx context.Context, pro
 
 			sessionComment.Attachments = append(sessionComment.Attachments, attachment)
 		} else if *s == modelInputs.IntegrationTypeJira {
-			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueProjectID, *issueTypeID); err != nil {
+			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueTeamID); err != nil {
 				return nil, e.Wrap(err, "error creating Jira task")
 			}
 
@@ -1887,7 +1886,7 @@ func (r *mutationResolver) ReplyToSessionComment(ctx context.Context, commentID 
 }
 
 // CreateErrorComment is the resolver for the createErrorComment field.
-func (r *mutationResolver) CreateErrorComment(ctx context.Context, projectID int, errorGroupSecureID string, text string, textForEmail string, taggedAdmins []*modelInputs.SanitizedAdminInput, taggedSlackUsers []*modelInputs.SanitizedSlackChannelInput, errorURL string, authorName string, issueTitle *string, issueDescription *string, issueTeamID *string, issueProjectID *string, issueTypeID *string, integrations []*modelInputs.IntegrationType) (*model.ErrorComment, error) {
+func (r *mutationResolver) CreateErrorComment(ctx context.Context, projectID int, errorGroupSecureID string, text string, textForEmail string, taggedAdmins []*modelInputs.SanitizedAdminInput, taggedSlackUsers []*modelInputs.SanitizedSlackChannelInput, errorURL string, authorName string, issueTitle *string, issueDescription *string, issueTeamID *string, integrations []*modelInputs.IntegrationType) (*model.ErrorComment, error) {
 	admin, isGuest := r.getCurrentAdminOrGuest(ctx)
 
 	errorGroup, err := r.canAdminViewErrorGroup(ctx, errorGroupSecureID)
@@ -2023,7 +2022,7 @@ func (r *mutationResolver) CreateErrorComment(ctx context.Context, projectID int
 
 			errorComment.Attachments = append(errorComment.Attachments, attachment)
 		} else if *s == modelInputs.IntegrationTypeJira {
-			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueProjectID, *issueTypeID); err != nil {
+			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueTeamID); err != nil {
 				return nil, e.Wrap(err, "error creating Jira task")
 			}
 
@@ -2128,7 +2127,7 @@ func (r *mutationResolver) MuteErrorCommentThread(ctx context.Context, id int, h
 }
 
 // CreateIssueForErrorComment is the resolver for the createIssueForErrorComment field.
-func (r *mutationResolver) CreateIssueForErrorComment(ctx context.Context, projectID int, errorURL string, errorCommentID int, authorName string, textForAttachment string, issueTitle *string, issueDescription *string, issueTeamID *string, issueProjectID *string, issueTypeID *string, integrations []*modelInputs.IntegrationType) (*model.ErrorComment, error) {
+func (r *mutationResolver) CreateIssueForErrorComment(ctx context.Context, projectID int, errorURL string, errorCommentID int, authorName string, textForAttachment string, issueTitle *string, issueDescription *string, issueTeamID *string, integrations []*modelInputs.IntegrationType) (*model.ErrorComment, error) {
 	var project model.Project
 	if err := r.DB.Where(&model.Project{Model: model.Model{ID: projectID}}).Take(&project).Error; err != nil {
 		return nil, err
@@ -2200,7 +2199,7 @@ func (r *mutationResolver) CreateIssueForErrorComment(ctx context.Context, proje
 
 			errorComment.Attachments = append(errorComment.Attachments, attachment)
 		} else if *s == modelInputs.IntegrationTypeJira {
-			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueProjectID, *issueTypeID); err != nil {
+			if err := r.CreateJiraTaskAndAttachment(ctx, workspace, attachment, title, desc, *issueTeamID); err != nil {
 				return nil, e.Wrap(err, "error creating Jira task")
 			}
