+++ frontend/src/graph/operators/mutation.gql
@@ -373,8 +373,6 @@ mutation CreateSessionComment(
 	$issue_team_id: String
 	$issue_description: String
 	$additional_context: String
-	$issue_type_id: String
-	$issue_project_id: String
 ) {
 	createSessionComment(
 		project_id: $project_id
@@ -396,8 +394,6 @@ mutation CreateSessionComment(
 		issue_team_id: $issue_team_id
 		issue_description: $issue_description
 		additional_context: $additional_context
-		issue_type_id: $issue_type_id
-		issue_project_id: $issue_project_id
 	) {
 		id
 		timestamp
@@ -431,8 +427,6 @@ mutation CreateIssueForSessionComment(
 	$issue_title: String
 	$issue_team_id: String
 	$issue_description: String
-	$issue_project_id: String
-	$issue_type_id: String
 ) {
 	createIssueForSessionComment(
 		project_id: $project_id
@@ -445,8 +439,6 @@ mutation CreateIssueForSessionComment(
 		issue_description: $issue_description
 		issue_team_id: $issue_team_id
 		integrations: $integrations
-		issue_project_id: $issue_project_id
-		issue_type_id: $issue_type_id
 	) {
 		id
 		timestamp
@@ -514,8 +506,6 @@ mutation CreateErrorComment(
 	$integrations: [IntegrationType]!
 	$issue_title: String
 	$issue_team_id: String
-	$issue_project_id: String
-	$issue_type_id: String
 	$issue_description: String
 ) {
 	createErrorComment(
@@ -530,8 +520,6 @@ mutation CreateErrorComment(
 		integrations: $integrations
 		issue_title: $issue_title
 		issue_team_id: $issue_team_id
-		issue_project_id: $issue_project_id
-		issue_type_id: $issue_type_id
 		issue_description: $issue_description
 	) {
 		id
@@ -555,8 +543,6 @@ mutation CreateIssueForErrorComment(
 	$issue_title: String
 	$issue_team_id: String
 	$issue_description: String
-	$issue_project_id: String
-	$issue_type_id: String
 ) {
 	createIssueForErrorComment(
 		project_id: $project_id
@@ -568,8 +554,6 @@ mutation CreateIssueForErrorComment(
 		issue_team_id: $issue_team_id
 		issue_description: $issue_description
 		integrations: $integrations
-		issue_project_id: $issue_project_id
-		issue_type_id: $issue_type_id
 	) {
 		id
 		created_at
