+++ frontend/src/graph/operators/query.gql
@@ -1553,11 +1553,6 @@ query GetJiraIntegrationSettings($workspace_id: ID!) {
 		id
 		name
 		key
-		issueTypes {
-			id
-			name
-			description
-		}
 	}
 }
