+++ frontend/src/pages/IntegrationsPage/IssueTrackerIntegrations.ts
@@ -11,8 +11,6 @@ import React from 'react'
 
 export interface ContainerSelectionProps {
 	setSelectionId: (id: string) => void
-	setIssueTypeId?: (id: string) => void
-	setIssueProjectId?: (id: string) => void
 	disabled: boolean
 }
